reorganized file structure to follow the atomic design architecture style, componentized nav bar buttons
