+++ src/components/Atoms/Buttons/NavServices.js
@@ -1,0 +1,9 @@
+import { style } from "../../../Style/Style.js";
+
+export default function NavBarButton() {
+  return (
+    <a style={style.navbarText} href="/services">
+      Services
+    </a>
+  );
+}
