--- src/components/HomepageComponents/Testimonials.js
@@ -1,7 +1,0 @@
-export default function Testimonials() {
-  return (
-    <section>
-        
-    </section>
-  )
-}
