+++ src/components/Molecules/NavBar/NavBar.js
@@ -1,0 +1,18 @@
+import { style } from "../../../Style/Style.js";
+import NavHome from "../../Atoms/Buttons/NavHome";
+import NavAbout from "../../Atoms/Buttons/NavAbout";
+import NavServices from "../../Atoms/Buttons/NavServices";
+import NavConnect from "../../Atoms/Buttons/NavConnect";
+
+export default function NavBar() {
+  return (
+    <nav style={style.navbarContainer}>
+      <div style={style.navbar}>
+        <NavHome />
+        <NavAbout />
+        <NavServices />
+        <NavConnect />
+      </div>
+    </nav>
+  );
+}
