+++ src/components/Organisms/Testimonials.js
@@ -1,0 +1,10 @@
+import { style } from "../../Style/Style.js"
+
+export default function Testimonials() {
+  return (
+    <section style={style.testimonialContainer}>
+        <div>
+        </div>
+    </section>
+  )
+}
