+++ src/pages/Homepage.js
@@ -1,4 +1,4 @@
-import HomeHeader from "../components/HomepageComponents/Header"
+import HomeHeader from "../components/Organisms/Header"
 
 export default function Homepage() {
   return <HomeHeader />
