+++ src/pages/Services.js
@@ -1,5 +1,5 @@
 export default function Services() {
   return (
-    <div>Soy Sauce</div>
+    <div>Services</div>
   )
 }
